pyproject.toml
[tool.poetry]
name = "azure-practice"
version = "0.1.0"
description = ""
authors = ["harisyammnv <harisyam.bphc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
azure-mgmt-resource = "^22.0.0"
azure-mgmt-datafactory = "^3.1.0"
azure-identity = "^1.12.0"

[tool.poetry.dev-dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
